warning
[git-p4: depot-paths = "//depot/qt/main/": change = 195654]

+++ src/gui/kernel/qeventdispatcher_x11.cpp
@@ -107,7 +107,7 @@ bool QEventDispatcherX11::processEvents(QEventLoop::ProcessEventsFlags flags)
     if (!d->interrupt) {
         const uint exclude_all =
             QEventLoop::ExcludeSocketNotifiers | QEventLoop::X11ExcludeTimers | QEventLoop::WaitForMoreEvents;
-        if (nevents > 0 && (flags & exclude_all) == exclude_all) {
+        if (nevents > 0 && ((uint)flags & exclude_all) == exclude_all) {
             QApplication::sendPostedEvents();
             return nevents > 0;
         }
